+++ packages/magic-products/src/min-spanning-trees/EdgeFrequencyDisplay.vue
